@@ -22,61 +22,62 @@
   },
   "private": true,
   "devDependencies": {
-    "babel-core": "^6.18.2",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^20.0.3",
-    "babel-loader": "^6.2.8",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-2": "^6.18.0",
-    "enzyme": "^2.9.1",
-    "eslint": "^4.6.1",
-    "eslint-config-airbnb": "^15.1.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-node": "^5.1.0",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^7.3.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "eslint-plugin-testcafe": "^0.2.1",
-    "html-webpack-plugin": "^2.24.1",
-    "jest": "^20.0.4",
-    "react": "^15.6.0",
-    "react-dom": "^15.6.0",
-    "react-test-renderer": "^15.6.1",
-    "testcafe": "^0.18.0",
+    "babel-core": "6.26.3",
+    "babel-eslint": "7.2.3",
+    "babel-jest": "20.0.3",
+    "babel-loader": "6.2.8",
+    "babel-preset-es2015": "6.18.0",
+    "babel-preset-react": "6.16.0",
+    "babel-preset-stage-2": "6.18.0",
+    "enzyme": "^3.6.0",
+    "eslint": "4.19.1",
+    "eslint-config-airbnb": "15.1.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-node": "5.2.1",
+    "eslint-plugin-promise": "3.8.0",
+    "eslint-plugin-react": "7.11.1",
+    "eslint-plugin-standard": "3.1.0",
+    "eslint-plugin-testcafe": "0.2.1",
+    "html-webpack-plugin": "2.24.1",
+    "jest": "20.0.4",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "react-test-renderer": "15.6.1",
+    "testcafe": "0.18.6",
     "testcafe-react-selectors": "0.0.8-alpha3",
-    "transfer-webpack-plugin": "^0.1.4",
+    "transfer-webpack-plugin": "0.1.4",
     "uglifyjs-webpack-plugin": "v1.0.0-beta.2",
-    "webpack": "^2.4.1",
-    "webpack-dev-server": "^2.4.2"
+    "webpack": "2.4.1",
+    "webpack-dev-server": "2.11.3"
   },
   "dependencies": {
-    "axios": "^0.16.2",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-es2017": "^6.24.1",
-    "body-parser": "^1.17.2",
-    "connect-redis": "^3.3.0",
-    "crypto-js": "^3.1.9-1",
-    "express": "^4.15.3",
-    "express-http-proxy": "^1.0.3",
-    "express-session": "^1.15.3",
-    "jsonminify": "^0.4.1",
-    "material-ui": "^0.18.2",
-    "prop-types": "^15.5.10",
-    "react": "^15.6.0",
-    "react-dom": "^15.6.0",
-    "react-highlight": "^0.10.0",
-    "react-lazylog": "^1.1.7",
-    "react-router-dom": "4.1.1",
-    "react-tap-event-plugin": "^2.0.1",
-    "recharts": "^1.0.0-alpha.4",
-    "request": "^2.81.0",
-    "transfer-webpack-plugin": "^0.1.4",
+    "@material-ui/codemod": "^1.1.0",
+    "@material-ui/core": "^3.1.0",
+    "axios": "0.16.2",
+    "babel-polyfill": "6.26.0",
+    "babel-preset-es2017": "6.24.1",
+    "body-parser": "1.18.3",
+    "connect-redis": "3.3.3",
+    "crypto-js": "3.1.9-1",
+    "express": "4.16.3",
+    "express-http-proxy": "1.2.0",
+    "express-session": "1.15.3",
+    "jsonminify": "0.4.1",
+    "material-ui": "0.18.2",
+    "prop-types": "15.6.2",
+    "react-event-listener": "^0.6.3",
+    "react-highlight": "^0.12.0",
+    "react-lazylog": "^3.1.4",
+    "react-router-dom": "^4.3.1",
+    "recharts": "1.2.0",
+    "recompose": "^0.30.0",
+    "request": "2.88.0",
+    "transfer-webpack-plugin": "0.1.4",
     "uglifyjs-webpack-plugin": "v1.0.0-beta.2",
-    "webpack": "^2.4.1",
-    "webpack-dev-middleware": "^1.10.2",
-    "webpack-hot-middleware": "^2.18.0"
+    "webpack": "2.4.1",
+    "webpack-dev-middleware": "1.12.2",
+    "webpack-hot-middleware": "2.24.0"
   }
 }
