@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Step, Stepper, StepLabel } from 'material-ui/Stepper';
-import RaisedButton from 'material-ui/RaisedButton';
+import Button from '@material-ui/core/Button';
 import FlatButton from 'material-ui/FlatButton';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
@@ -180,15 +180,16 @@ export default class AttachAddon extends Component {
               className="back"
               label="Back"
               disabled={stepIndex === 0}
-              onTouchTap={this.handlePrev}
+              onClick={this.handlePrev}
               style={style.buttons.Back}
             />)}
             {stepIndex > 0 && (
-              <RaisedButton
+              <Button
+                variant="contained"
                 className="next"
                 label={stepIndex === 1 ? 'Finish' : 'Next'}
                 primary
-                onTouchTap={this.handleNext}
+                onClick={this.handleNext}
               />
             )}
           </div>
@@ -225,7 +226,7 @@ export default class AttachAddon extends Component {
                 className="ok"
                 label="Ok"
                 primary
-                onTouchTap={this.handleClose}
+                onClick={this.handleClose}
               />}
           >
             {this.state.submitMessage}
