@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Step, Stepper, StepLabel } from 'material-ui/Stepper';
 import { RadioButton, RadioButtonGroup } from 'material-ui/RadioButton';
-import RaisedButton from 'material-ui/RaisedButton';
+import Button from '@material-ui/core/Button';
 import FlatButton from 'material-ui/FlatButton';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
@@ -257,14 +257,15 @@ export default class NewFormation extends Component {
             className="back"
             label="Back"
             disabled={stepIndex === 0}
-            onTouchTap={this.handlePrev}
+            onClick={this.handlePrev}
             style={style.buttons.back}
           />)}
-          <RaisedButton
+          <Button
+            variant="contained"
             className="next"
             label={stepIndex === 3 ? 'Finish' : 'Next'}
             primary
-            onTouchTap={this.handleNext}
+            onClick={this.handleNext}
           />
         </div>
       </div>
@@ -304,7 +305,7 @@ export default class NewFormation extends Component {
                 className="ok"
                 label="Ok"
                 primary
-                onTouchTap={this.handleClose}
+                onClick={this.handleClose}
               />}
           >
             {this.state.submitMessage}
